Synching improved, but may cause crasches. Single player not affected. Jumping works on both server and client side now.

+++ Circle.cpp
@@ -10,7 +10,7 @@ Shape(game, position, dynamic, density, friction, groupIndex, id)
     
     // Define another box shape for our dynamic body.
     b2CircleShape circleShape; //dynamicBox;
-    circleShape.m_radius = radius * game->getUtility()->getPTM();
+    circleShape.m_radius = radius * (float)game->getUtility()->getPTM();
 
 //    boxShape.SetAsBox((size->x/2) * pixelToMeter, (size->y/2) * pixelToMeter);
 //    circleS
